delete the tag item when delete the issue associated

+++ issue-service/src/main/java/cn/edu/fudan/issueservice/service/impl/IssueServiceImpl.java
@@ -84,7 +84,14 @@ public class IssueServiceImpl implements IssueService {
 
     @Override
     public void deleteIssueByProjectId(String projectId) {
+        //先删除该项目所有issue对应的tag
+        List<String> issueIds=issueDao.getIssueIdsByProjectId(projectId);
+        JSONObject response=restTemplate.postForObject(tagServicePath+"/tagged-delete",issueIds,JSONObject.class);
+        if(response==null||response.getIntValue("code")!=200){
+            throw new RuntimeException("tag item delete failed!");
+        }
         issueDao.deleteIssueByProjectId(projectId);
+
     }
 
     @Override
@@ -220,6 +227,11 @@ public class IssueServiceImpl implements IssueService {
 //        }
         result.put("data",getFakeData());
         return result;
+    }
+
+    @Override
+    public Object getExistIssueTypes() {
+        return issueDao.getExistIssueTypes();
     }
 
     @SuppressWarnings("unchecked")
